+++ src/helix/endpoints/moderation/mod.rs
@@ -8,6 +8,7 @@ use crate::{
 use serde::{Deserialize, Serialize};
 
 pub mod add_blocked_term;
+pub mod add_channel_moderator;
 pub mod ban_user;
 pub mod check_automod_status;
 pub mod delete_chat_messages;
@@ -16,10 +17,13 @@ pub mod get_blocked_terms;
 pub mod get_moderators;
 pub mod manage_held_automod_messages;
 pub mod remove_blocked_term;
+pub mod remove_channel_moderator;
 pub mod unban_user;
 
 #[doc(inline)]
 pub use add_blocked_term::{AddBlockedTermBody, AddBlockedTermRequest};
+#[doc(inline)]
+pub use add_channel_moderator::{AddChannelModeratorRequest, AddChannelModeratorResponse};
 #[doc(inline)]
 pub use ban_user::{BanUser, BanUserBody, BanUserRequest};
 #[doc(inline)]
@@ -39,6 +43,8 @@ pub use manage_held_automod_messages::{
 };
 #[doc(inline)]
 pub use remove_blocked_term::{RemoveBlockedTerm, RemoveBlockedTermRequest};
+#[doc(inline)]
+pub use remove_channel_moderator::{RemoveChannelModeratorRequest, RemoveChannelModeratorResponse};
 #[doc(inline)]
 pub use unban_user::{UnbanUserRequest, UnbanUserResponse};