+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Added `Send Chat Announcement` endpoint
 * Added `Delete Chat Messages` endpoint
 * Added `Get User Chat Color` and `Update User Chat Colors` endpoint
+* Added `Add Channel Moderator` endpoint
 
 ## [v0.6.1] - 2022-04-29